Add esutil package
The complication is that esutil ships with its own `ups/esutil.table` file.
It also overwrites the shipped version in `setup.py`.

We want to use our own custom table file.  To do this I did the following:
* made an override to `prep` to link the LSST version into place after the package
is unpacked.
* added a patch that removes the parts of `setup.py` that rewrite the file.

Note that we still need to include a link to the LSST table file in the repo, even though
it is overwritten, to setup the environment initially.